requirements.txt
aiohttp==3.8.3
aiosignal==1.2.0
APScheduler==3.6.3
async-timeout==4.0.2
attrs==22.1.0
certifi==2022.9.24
charset-normalizer==2.1.1
frozenlist==1.3.1
idna==3.4
multidict==6.0.2
pyTelegramBotAPI==4.7.0
python-dotenv==1.0.0
pytz==2022.5
pytz-deprecation-shim==0.1.0.post0
requests==2.28.1
six==1.16.0
telebot==0.0.4
tornado==6.2
tzdata==2022.5
tzlocal==4.2
urllib3==1.26.12
yarl==1.8.1
